BO revamp
Curators separated of playlists
Possibility to edit the playlist fields
Possibility to work on future playlists

+++ control/delete_track.php
@@ -1,5 +1,13 @@
 <?php
-	include_once("connect_sql.php");
-	$req = $bdd->prepare('DELETE FROM song WHERE id = ?');
-	$req->execute(array($_GET['idSong']));
-	header('Location: http://soundpark.fm/view/create_playlist.php');
+	include_once("../model/connect_sql.php");
+	if(isset($_GET['idSong']))
+	{
+		$req = $bdd->prepare('DELETE FROM song WHERE id = ?');
+		$req->execute(array($_GET['idSong']));
+		header('Location: http://soundpark.fm/view/create_playlist.php');
+	}
+	else
+	{
+		echo('No songId');
+	}
+
